services/correction-serv/requirements.txt
# ===============================
# Core API
# ===============================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ===============================
# Database
# ===============================
motor==3.3.2
python-dotenv==1.0.1

# ===============================
# Rules & Validation
# ===============================
pyyaml==6.0.1
pydantic==2.6.4

# ===============================
# Machine Learning - Lightweight
# ===============================
numpy==1.26.4
scikit-learn==1.4.1.post1

# ===============================
# NLP - Text Correction (T5)
# ===============================
torch==2.2.1
transformers==4.39.3
sentencepiece==0.2.0

# ===============================
# Observability / Utils
# ===============================
loguru==0.7.2
